[AC-5039] move models/utils.py to models_utils.py
